[FEATURE] More validations on YAML rule editor (#279)
* yaml editor as custom form component


* refactor file structure for rule editor


* rename RuleEditorFormState to RuleEditorFormModel


* refactor yaml rule editor custom form component


* pr comments fixes and snapshot tests


* more snapshot tests


* align fields labels styling with create detector


* aligne formik version with alerting plugin


* addresing validation UX issues


* validate tag format
