+++ package.json
@@ -67,6 +67,6 @@
     "yarn": "^1.21.1"
   },
   "dependencies": {
-    "formik": "^2.2.9"
+    "formik": "^2.2.6"
   }
 }
