+++ public/pages/Rules/containers/ImportRule/ImportRule.tsx
@@ -4,7 +4,7 @@
  */
 
 import { BrowserServices } from '../../../../models/interfaces';
-import { RuleEditor } from '../../components/RuleEditor/RuleEditor';
+import { RuleEditorContainer } from '../../components/RuleEditor/RuleEditorContainer';
 import React, { useCallback, useContext, useEffect, useState } from 'react';
 import { EuiButton, EuiFilePicker, EuiFlexGroup, EuiFlexItem, EuiSpacer } from '@elastic/eui';
 import { BREADCRUMBS, ROUTES } from '../../../../utils/constants';
@@ -72,7 +72,14 @@ export const ImportRule: React.FC<ImportRuleProps> = ({ history, services, notif
               })) || [],
           };
           setContent(
-            <RuleEditor title="Import a rule" FooterActions={footerActions} rule={rule} />
+            <RuleEditorContainer
+              title="Import a rule"
+              history={history}
+              notifications={notifications}
+              mode={'create'}
+              ruleService={services.ruleService}
+              rule={rule}
+            />
           );
         } catch (error: any) {
           setFileError('Invalid file content');
